+++ CHANGELOG.md
@@ -2,14 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-### [0.10.1](https://github.com/amplience/dc-demostore-cli/compare/v0.10.0...v0.10.1) (2022-03-24)
+### [1.1.2](https://github.com/amplience/dc-demostore-cli/compare/v1.2.0...v1.1.2) (2022-06-22)
 
+### Bug Fixes
+
+* added validation level before creation ([0aaf3d3](https://github.com/amplience/dc-demostore-cli/commit/0aaf3d39b3dfb848cf5f1bdbfc29cf4dea2023c7))
+* validation level on codecs schema ([82a1c05](https://github.com/amplience/dc-demostore-cli/commit/82a1c05778532590098bfccff7dcc607d01d8fa9))
+
+### [0.10.1](https://github.com/amplience/dc-demostore-cli/compare/v0.10.0...v0.10.1) (2022-03-24)
 
 ### Bug Fixes
 
 * remove prepare life cycle script to avoid building during npx run which is unnecessary because the npm package contains the dist js files ([6206c87](https://github.com/amplience/dc-demostore-cli/commit/6206c87fc94580bf515504fc19bdaa8ec2660b6b))
 
-## [0.10.0](https://github.com/amplience/dc-demostore-cli/compare/v0.9.1...v0.10.0) (2022-03-24)
+### [0.10.0](https://github.com/amplience/dc-demostore-cli/compare/v0.9.1...v0.10.0) (2022-03-24)
 
 ### [0.9.1](https://github.com/amplience/dc-demostore-cli/compare/v0.2.22...v0.9.1) (2022-03-23)